ir: Move clustering Queue to processing.py
